requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.0.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
markdown>=3.5.0
weasyprint>=61.0
